modified structure of UI
